refactor: useTimerPageModal의 UI로직을 별도의 컴포넌트로 분리
